requirements.txt
appdirs==1.4.0
Django==1.10.5
django-cors-headers==2.0.1
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
google-api-python-client==1.6.1
httplib2==0.9.2
oauth2client==4.0.0
packaging==16.8
pyasn1==0.1.9
pyasn1-modules==0.0.8
PyJWT==1.4.2
pyparsing==2.1.10
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
